refactor: Folder structure and sort classes in subfolders
This prepares to split up our too large
classes into smaller once with
extensions.
