+++ lib/matrix.dart
@@ -21,17 +21,17 @@ library matrix;
 
 export 'package:matrix_api_lite/matrix_api_lite.dart';
 
-export 'src/client.dart';
+export 'src/client/client.dart';
 export 'src/database/database_api.dart';
 export 'src/database/hive_database.dart';
 export 'src/database/fluffybox_database.dart';
-export 'src/event.dart';
-export 'src/event_status.dart';
-export 'src/voip.dart';
-export 'src/voip_content.dart';
-export 'src/room.dart';
-export 'src/timeline.dart';
-export 'src/user.dart';
+export 'src/event/event.dart';
+export 'src/event/event_status.dart';
+export 'src/voip/voip.dart';
+export 'src/voip/voip_content.dart';
+export 'src/room/room.dart';
+export 'src/room/timeline.dart';
+export 'src/event/user.dart';
 export 'src/utils/commands_extension.dart';
 export 'src/utils/crypto/encrypted_file.dart';
 export 'src/utils/device_keys_list.dart';
@@ -46,4 +46,4 @@ export 'src/utils/sync_update_extension.dart';
 export 'src/utils/to_device_event.dart';
 export 'src/utils/uia_request.dart';
 export 'src/utils/uri_extension.dart';
-export 'src/voip_content.dart';
+export 'src/voip/voip_content.dart';
